Move P2pService to its own module
