+++ network-grpc/src/client.rs
@@ -9,7 +9,10 @@ use crate::{
 };
 
 use chain_core::property;
-use network_core::client::{Client, block::BlockService, gossip::GossipService, P2pService};
+use network_core::client::Client;
+use network_core::client::block::BlockService;
+use network_core::client::gossip::GossipService;
+use network_core::client::p2p::P2pService;
 use network_core::error as core_error;
 use network_core::gossip::{self, Gossip, NodeId};
 use network_core::subscription::BlockEvent;
